[polling] collect order books
